Re-factored hosts list data state
Now the data that is used in the host list is directly get from the
redux store using a selector created with Redux Toolkit (memoized input selectors + result
combinator). Also, moved the Cidr calculation function into @lib
