+++ assets/js/lib/ip/index.js
@@ -1,0 +1,12 @@
+// SPDX-FileCopyrightText: SUSE LLC
+// SPDX-License-Identifier: Apache-2.0
+
+import { zipWith } from 'lodash';
+
+export function buildCidrNotation(ipAddresses, netmasks) {
+  return zipWith(
+    ipAddresses,
+    netmasks,
+    (address, netmask) => `${address}${netmask ? `/${netmask}` : ''}`
+  );
+}
