+++ assets/js/pages/HostsList/HostsList.jsx
@@ -25,6 +25,16 @@ import { post, del } from '@lib/network';
 import { agentVersionWarning } from '@lib/agent';
 import { STALE_ROW } from '@lib/tables';
 
+import ClusterLink from '@pages/ClusterDetails/ClusterLink';
+import DeregistrationModal from '@pages/DeregistrationModal';
+import HealthSummary from '@pages/HealthSummary';
+import { getCounters } from '@pages/HealthSummary/summarySelection';
+
+import { addTagToHost, removeTagFromHost, deregisterHost } from '@state/hosts';
+import { getUserProfile } from '@state/selectors/user';
+import { hostsDataSelector } from '@state/selectors/host';
+import { getInstanceID } from '@state/instances';
+
 const compareAgentVersions = (a, b) => {
   const coercedA = semver.coerce(a);
   const coercedB = semver.coerce(b);
@@ -33,20 +43,6 @@ const compareAgentVersions = (a, b) => {
   }
   return String(a).localeCompare(String(b));
 };
-
-import ClusterLink from '@pages/ClusterDetails/ClusterLink';
-import DeregistrationModal from '@pages/DeregistrationModal';
-import HealthSummary from '@pages/HealthSummary';
-import { getCounters } from '@pages/HealthSummary/summarySelection';
-import { buildCidrNotation } from '@pages/HostDetailsPage/HostDetails';
-
-import { addTagToHost, removeTagFromHost, deregisterHost } from '@state/hosts';
-import { getAllSAPInstances } from '@state/selectors/sapSystem';
-import { getUserProfile } from '@state/selectors/user';
-import { getInstanceID } from '@state/instances';
-
-const getInstancesByHost = (instances, hostId) =>
-  instances.filter((instance) => instance.host_id === hostId);
 
 const addTag = (tag, hostId) => {
   post(`/hosts/${hostId}/tags`, {
@@ -59,9 +55,7 @@ const removeTag = (tag, hostId) => {
 };
 
 function HostsList() {
-  const hosts = useSelector((state) => state.hostsList.hosts);
-  const clusters = useSelector((state) => state.clustersList.clusters);
-  const allInstances = useSelector(getAllSAPInstances);
+  const hostsData = useSelector(hostsDataSelector);
   const { abilities, timezone } = useSelector(getUserProfile);
 
   const [searchParams, setSearchParams] = useSearchParams();
@@ -79,6 +73,8 @@ function HostsList() {
     setCleanUpModalOpen(false);
     dispatch(deregisterHost({ id, hostname }));
   };
+
+  const counters = getCounters(hostsData || []);
 
   const config = {
     pagination: true,
@@ -243,28 +239,6 @@ function HostsList() {
     ],
   };
 
-  const data = hosts.map((host) => {
-    const cluster = clusters.find((c) => c.id === host.cluster_id);
-    const sapSystemList = getInstancesByHost(allInstances, host.id);
-
-    return {
-      health: host.health,
-      hostname: host.hostname,
-      ip: buildCidrNotation(host.ip_addresses, host.netmasks),
-      provider: host.provider,
-      sid: sapSystemList.map((sapSystem) => sapSystem.sid),
-      cluster,
-      agent_version: host.agent_version,
-      id: host.id,
-      tags: (host.tags && host.tags.map((tag) => tag.value)) || [],
-      sap_systems: sapSystemList,
-      deregisterable: host.deregisterable,
-      deregistering: host.deregistering,
-      staleAt: host.stale_at,
-    };
-  });
-
-  const counters = getCounters(data || []);
   return (
     <>
       <PageHeader className="font-bold">Hosts</PageHeader>
@@ -282,7 +256,7 @@ function HostsList() {
         <HealthSummary {...counters} className="px-4 py-2" />
         <Table
           config={config}
-          data={data}
+          data={hostsData}
           searchParams={searchParams}
           setSearchParams={setSearchParams}
         />
